# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "description": "API mô tả tài liệu cho hệ thống Laravel",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/test/{index2}": {
            "get": {
                "tags": [
                    "Test"
                ],
                "summary": "Swagger App\\Http\\Controllers\\TestController@index",
                "description": "",
                "operationId": "50aae5c297df7631e1623b93b708549e",
                "parameters": [
                    {
                        "name": "index2",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "classId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "keyWork",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "integer"
                                        },
                                        "email": {
                                            "type": "integer"
                                        },
                                        "classes": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "host": [],
                        "user-agent": [],
                        "accept": [],
                        "accept-language": [],
                        "accept-charset": []
                    }
                ]
            }
        },
        "/api/test-2": {
            "get": {
                "tags": [
                    "Test"
                ],
                "summary": "Swagger App\\Http\\Controllers\\TestController@index2",
                "description": "",
                "operationId": "ad44ad58379dbfd5c6f6bf312b0e76d1",
                "responses": {
                    "200": {
                        "description": "Return information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "phone": {
                                            "type": "string"
                                        },
                                        "phone2": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "host": [],
                        "user-agent": [],
                        "accept": [],
                        "accept-language": [],
                        "accept-charset": []
                    }
                ]
            }
        },
        "/api/test-3": {
            "get": {
                "tags": [
                    "Test"
                ],
                "summary": "Swagger App\\Http\\Controllers\\TestController@index3",
                "description": "",
                "operationId": "2dee3ea3e391288902fdfb1e7e7ec4be",
                "responses": {
                    "200": {
                        "description": "Return information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "phone": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "host": [],
                        "user-agent": [],
                        "accept": [],
                        "accept-language": [],
                        "accept-charset": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Test",
            "description": "Test"
        }
    ]
}